Fixed player creation and connect with new entity script API.
Network entity map is now a global variable. This is so that network components can add entities to the network entity map as they are created from a script (no delay waiting for a registration system). Plus, there should be only one network entity map in an application anyway. Tests have been updated, are a bit wonky, but works well within the application.